+++ homework4/file_search/frequencySearch.h
@@ -1,8 +1,4 @@
 #pragma once
 
-#include <stdio.h>
-
 int frequencySearch(int *sourceArray, int arrayLength);
 void quickSort(int *qsSourceArray, int sortLeftBorder, int sortRightBorder);
-bool programTest();
-bool scanFile(FILE *inputFile, int *array, int arrayLength);
