+++ homework4/file_search/scanFile.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <stdio.h>
+
+bool scanFile(FILE *inputFile, int *array, int arrayLength);
